% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc68k}
\alias{pbmc68k}
\title{A matrix of the top 50 PCs of part of the pbmc68k dataset.}
\format{
A matrix of 2127 cells and 50 PCs
}
\usage{
data("pbmc68k")
}
\description{
A matrix of the top 50 PCs of part of the pbmc68k dataset.
}
\keyword{datasets}
