% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scsampler}
\alias{scsampler}
\title{The scSampler function}
\usage{
scsampler(
  dat,
  use_pca = TRUE,
  run_pca = TRUE,
  use_dim = 50,
  assay_use = "logcounts",
  fraction = 0.1,
  n_obs = NULL,
  random_state = 1,
  copy = FALSE,
  random_split = 1
)
}
\arguments{
\item{dat}{The input single cell dataset. Can be one of: a \code{SingleCellExperiment} object, a \code{SeuratObject} object, or a matrix.}

\item{use_pca}{A logical value. If TRUE, the algorithm will use the PC space for subsampling. We highly recommend setting it as TRUE.}

\item{run_pca}{A logical value. If TRUE, run PCA by \code{irlba::prcomp_irlba}. If FALSE, use the PCs from the input object.}

\item{use_dim}{An integer of how many PCs are used. Defalut is 50.}

\item{assay_use}{The input expression space.}

\item{fraction}{A numeric value from 0 to 1. The fraction of the subsample.}

\item{n_obs}{An integer of the size of the subsample.}

\item{random_state}{An integer of the random seed.}

\item{copy}{A logical value. If TRUE, return a subset of the origianl object, otherwise the indices of the subsample. Default is FALSE.}

\item{random_split}{An integer of how many splits for speeding up. Default is 1 (no split).}
}
\value{
A subset of the original object or the subsample indicies.
}
\description{
\code{scSampler} takes the input single cell data and return the selected diverse subsample.
}
\examples{
data("pbmc68k")
res <- scsampler(dat = pbmc68k, use_pca = FALSE, run_pca = FALSE)
}
